feat: add base form input
